feat: Added editor test project.
