+++ EditorTests/Ludus/EditorTests/SmokeTests.cpp
@@ -1,0 +1,9 @@
+#include "pch.h"
+
+namespace Ludus::EditorTests
+{
+	TEST(SmokeTests, Compiles)
+	{
+		SUCCEED();
+	}
+}
