+++ EngineTests/Ludus/EngineTests/SmokeTests.cpp
@@ -1,0 +1,9 @@
+#include "pch.h"
+
+namespace Ludus::EngineTests
+{
+	TEST(SmokeTests, Compiles)
+	{
+		SUCCEED();
+	}
+}
