+++ Ludus.sln
@@ -10,7 +10,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		THIRD_PARTY_NOTICES.md = THIRD_PARTY_NOTICES.md
 	EndProjectSection
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Tests", "Tests\Tests.vcxproj", "{C6D4E52B-8470-417D-96AD-48DBE33B43ED}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "EngineTests", "EngineTests\EngineTests.vcxproj", "{C6D4E52B-8470-417D-96AD-48DBE33B43ED}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "EditorTests", "EditorTests\EditorTests.vcxproj", "{5B01F46F-7980-4E1D-BC57-6633174CB035}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "UI", "UI\UI.vcxproj", "{D7F5C7FB-E1A5-4C5B-9BD2-EAA351CA76A2}"
 EndProject
@@ -30,6 +32,10 @@ Global
 		{C6D4E52B-8470-417D-96AD-48DBE33B43ED}.Debug|x64.Build.0 = Debug|x64
 		{C6D4E52B-8470-417D-96AD-48DBE33B43ED}.Release|x64.ActiveCfg = Release|x64
 		{C6D4E52B-8470-417D-96AD-48DBE33B43ED}.Release|x64.Build.0 = Release|x64
+		{5B01F46F-7980-4E1D-BC57-6633174CB035}.Debug|x64.ActiveCfg = Debug|x64
+		{5B01F46F-7980-4E1D-BC57-6633174CB035}.Debug|x64.Build.0 = Debug|x64
+		{5B01F46F-7980-4E1D-BC57-6633174CB035}.Release|x64.ActiveCfg = Release|x64
+		{5B01F46F-7980-4E1D-BC57-6633174CB035}.Release|x64.Build.0 = Release|x64
 		{D7F5C7FB-E1A5-4C5B-9BD2-EAA351CA76A2}.Debug|x64.ActiveCfg = Debug|x64
 		{D7F5C7FB-E1A5-4C5B-9BD2-EAA351CA76A2}.Debug|x64.Build.0 = Debug|x64
 		{D7F5C7FB-E1A5-4C5B-9BD2-EAA351CA76A2}.Release|x64.ActiveCfg = Release|x64
