--- Tests/Ludus/Tests/Serialization/Schemas/DisplayNameComponentSchemaTests.cpp
@@ -1,144 +1,0 @@
-#include "pch.h"
-
-#include <cstdint>
-#include <string>
-
-#include <Ludus/Engine/Components/DisplayNameComponent.h>
-#include <Ludus/Engine/Serialization/Core/DomDocument.h>
-#include <Ludus/Engine/Serialization/Core/DomNode.h>
-#include <Ludus/Engine/Serialization/Core/DomTokenStreamReader.h>
-#include <Ludus/Engine/Serialization/Core/DomTokenStreamWriter.h>
-#include <Ludus/Engine/Serialization/Core/Token.h>
-#include <Ludus/Engine/Serialization/Schemas/DisplayNameComponentSchema.h>
-
-namespace Ludus::Tests::Serialization::Schemas
-{
-	using DomDocument = Ludus::Engine::Serialization::Core::DomDocument;
-	using DomTokenStreamWriter = Ludus::Engine::Serialization::Core::DomTokenStreamWriter;
-	using DomTokenStreamReader = Ludus::Engine::Serialization::Core::DomTokenStreamReader;
-	using DisplayNameComponentSchema = Ludus::Engine::Serialization::Schemas::DisplayNameComponentSchema;
-	using Token = Ludus::Engine::Serialization::Core::Token;
-
-	using Ludus::Engine::Serialization::Core::AsObject;
-	using Ludus::Engine::Serialization::Core::AsValue;
-	using Ludus::Engine::Serialization::Core::DomNode;
-	using Ludus::Engine::Serialization::Core::DomObject;
-
-	static const DomNode* FindMember(const DomObject& object, std::string_view key)
-	{
-		for (const auto& [memberKey, memberValue] : object)
-		{
-			if (memberKey == key)
-			{
-				return memberValue.get();
-			}
-		}
-
-		return nullptr;
-	}
-
-	TEST(DisplayNameComponentSchema, Serialize_WritesExpectedComponentValues)
-	{
-		// Arrange.
-		Ludus::Engine::Components::DisplayNameComponent displayName(1, "Name");
-		DomDocument document;
-		DomTokenStreamWriter writer(document);
-
-		// Act.
-		DisplayNameComponentSchema::Serialize(writer, displayName);
-
-		// Assert.
-		const auto* root = document.GetRoot();
-		ASSERT_NE(root, nullptr);
-
-		const auto& displayNameObject = AsObject(*root);
-		const auto* ownerNode = FindMember(displayNameObject, "OwnerHandle");
-		const auto* valueNode = FindMember(displayNameObject, "Value");
-
-		ASSERT_NE(ownerNode, nullptr);
-		ASSERT_NE(valueNode, nullptr);
-
-		const auto ownerHandle = std::get<uint64_t>(AsValue(*ownerNode));
-		const auto value = std::get<std::string>(AsValue(*valueNode));
-
-		ASSERT_EQ(ownerHandle, 1u);
-		ASSERT_EQ(value, "Name");
-	}
-
-	TEST(DisplayNameComponentSchema, Deserialize_ReadsFields_When_ComponentHasValues)
-	{
-		// Arrange.
-		DomDocument document;
-		DomTokenStreamWriter writer(document);
-		writer.Emit(Token::StartObject { });
-		writer.Emit(Token::Key { "OwnerHandle" });
-		writer.Emit(Token::Int { 1 });
-		writer.Emit(Token::Key { "Value" });
-		writer.Emit(Token::String { "Name" });
-		writer.Emit(Token::EndObject { });
-		DomTokenStreamReader reader(document);
-
-		// Act.
-		const auto result = DisplayNameComponentSchema::Deserialize(reader);
-
-		// Assert.
-		ASSERT_TRUE(result.HasValue());
-
-		const auto& displayNameResult = result.GetValue();
-		ASSERT_EQ(displayNameResult.OwnerHandle, 1u);
-		ASSERT_EQ(displayNameResult.Value, "Name");
-	}
-
-	TEST(DisplayNameComponentSchema, Deserialize_DefaultsOptionalFields_When_Missing)
-	{
-		// Arrange.
-		DomDocument document;
-		DomTokenStreamWriter writer(document);
-		writer.Emit(Token::StartObject { });
-		writer.Emit(Token::Key { "OwnerHandle" });
-		writer.Emit(Token::Int { 1 });
-		writer.Emit(Token::EndObject { });
-		DomTokenStreamReader reader(document);
-
-		// Act.
-		const auto result = DisplayNameComponentSchema::Deserialize(reader);
-
-		// Assert.
-		ASSERT_TRUE(result.HasValue());
-
-		const auto& displayNameResult = result.GetValue();
-		ASSERT_EQ(displayNameResult.OwnerHandle, 1u);
-		ASSERT_EQ(displayNameResult.Value, "");
-	}
-
-	TEST(DisplayNameComponentSchema, Deserialize_Fails_When_RequiredFieldsAreMissing)
-	{
-		// Arrange.
-		DomDocument document;
-		DomTokenStreamWriter writer(document);
-		writer.Emit(Token::StartObject { });
-		writer.Emit(Token::EndObject { });
-		DomTokenStreamReader reader(document);
-
-		// Act.
-		const auto result = DisplayNameComponentSchema::Deserialize(reader);
-
-		// Assert.
-		ASSERT_FALSE(result.HasValue());
-	}
-
-	TEST(DisplayNameComponentSchema, Deserialize_Fails_When_ComponentHeaderIsMissing)
-	{
-		// Arrange.
-		DomDocument document;
-		DomTokenStreamWriter writer(document);
-		writer.Emit(Token::Int { 1 });
-		DomTokenStreamReader reader(document);
-
-		// Act.
-		const auto result = DisplayNameComponentSchema::Deserialize(reader);
-
-		// Assert.
-		ASSERT_FALSE(result.HasValue());
-	}
-}
